[BE]chore: 회원 대시보드 조회 기능 ing(연속학습일수 완성)

+++ backend/src/main/java/com/tunapearl/saturi/dto/user/UserContinuousLearnDayDTO.java
@@ -9,5 +9,5 @@ import java.util.List;
 @AllArgsConstructor
 public class UserContinuousLearnDayDTO {
     private Long learnDays;
-    private List<Long> daysOfTheWeek;
+    private List<Integer> daysOfTheWeek;
 }
